add choose_mode

+++ service.sh
@@ -9,7 +9,11 @@ until [ $(getprop sys.boot_completed) -eq 1 ] && [ -d "/sdcard/Documents" ]; do
     exit 0
   fi
 done
-rm -rf ${pid_file}
+
+if [ -f ${pid_file} ] ; then
+    rm -rf ${pid_file}
+fi
+
 clash_control enable
 # default disable ipv6 accept_ra
 echo 0 > /proc/sys/net/ipv6/conf/all/accept_ra
